+++ src/components/index.js
@@ -1,6 +1,6 @@
-import Header from './Header/Header.view';
+import Header from './Header/Header';
 import Card from './Card/Card.view';
-import FavoriteButton from './FavoriteButton/FavoriteButton.view';
-import Search from './Search/Search.view';
+import FavoriteButton from './FavoriteButton/FavoriteButton';
+import Search from './Search/Search';
 
 export { Search, Card, Header, FavoriteButton };
